requirements.txt
python-dotenv
# metagpt==0.7.4
openai
pdfminer.six
chromadb
nltk
jieba
elasticsearch7
sentence_transformers
bs4
qianfan
pytest
langchain
langchain_community
langgraph
langserve
langchain_openai
langchain_chroma
langchainhub
faiss-cpu
anyio
pytest-asyncio
google-search-results
